dayflow-hrms-backend/employees/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path, include
from .views import EmployeeViewSet, DepartmentViewSet

router = DefaultRouter()
router.register(r'list', EmployeeViewSet, basename='employee')
router.register(r'departments', DepartmentViewSet, basename='department')

urlpatterns = [
    path('', include(router.urls)),
]
