dayflow-hrms-backend/requirements.txt
Django>=4.2.7
djangorestframework>=3.14.0
rest_framework_simplejwt>=5.3.0
daphne>=4.0.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
drf-spectacular>=0.26.5
celery>=5.3.4
redis>=5.0.1
django-environ>=0.11.2
Pillow>=10.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
channels>=4.0.0
channels-redis>=4.1.0
argon2-cffi>=23.1.0
